+++ Tests/PropTests/JustifyEndTests.cs
@@ -1,5 +1,6 @@
 using Microsoft.Xna.Framework;
 using NUnit.Framework;
+using WForest.Exceptions;
 using WForest.Factories;
 using WForest.UI.Props.Grid.JustifyProps;
 using WForest.UI.Widgets.Interfaces;
@@ -27,16 +28,12 @@ namespace WForest.Tests.PropTests
         }
 
         [Test]
-        public void WidgetWithoutRowOrCol_NothingHappens()
+        public void WidgetWithoutRowOrCol_ThrowsExc()
         {
             var child = WidgetFactory.Container(new Rectangle(0, 0, 120, 120));
             _root.AddChild(child);
-
-            _justifyEnd.ApplyOn(_root);
-
-            var expected = new Rectangle(0, 0, 120, 120);
-
-            Assert.That(child.Space, Is.EqualTo(expected));
+            
+            Assert.That(() => _justifyEnd.ApplyOn(_root), Throws.TypeOf<IncompatibleWidgetException>());
         }
 
         [Test]